# Speed up Data.ByteString.Short.unpack