@Bodigrim 2022-07-03
Git history reveals that `Data.ByteString.Short.Internal.unpackAppend{Chars,Bytes}{Lazy,Strict}` were added in 9e2814ff65a51d0684eb1b416fad8bf809685faf, copied from `Data.ByteString.Internal.unpackAppend{Chars,Bytes}{Lazy,Strict}`. 

https://github.com/haskell/bytestring/blob/2b9416dd73121b5b2abd7a4594da249028858e4b/Data/ByteString/Internal.hs#L454-L457

https://github.com/haskell/bytestring/blob/2b9416dd73121b5b2abd7a4594da249028858e4b/Data/ByteString/Internal.hs#L494-L502

This is a reasonable strategy for `ByteString`s, because indexing them involves `accursedUnutterablePerformIO $ unsafeWithForeignPtr fp`, which is a non-negligible cost if performed per each element. The same does not hold for `ShortByteString`, because `indexWord8Array` is as cheap as it gets. Thanks for noticing this, @hasufell!