@hasufell 2022-07-03
My laptop might not be too reliable about performance tests, but I ran it twice and it seems that that `foldr` got consistently slower: https://gist.github.com/hasufell/33c0736e4fcdd1be4896c08322f191f1