@sjakobi 2022-07-04
Could you change `foldr'` so it operates on the `ByteArray#` directly instead of consuming it via `unpack`?

I don't find it very surprising that a `foldr'` that consumes a list isn't efficient.