@sjakobi 2022-07-04
Comparing the Core between the old and new versions, both for `foldr'` itself and the `foldr'` benchmark might reveal the issue.