@hasufell 2022-07-04
Even regular `foldr` is slower.

I was able to speed them up by writing explicit versions:

```hs
foldr :: (Word8 -> a -> a) -> a -> ShortByteString -> a
foldr k v = \sbs ->
  let l  = length sbs
      ba = asBA sbs
      w  = indexWord8Array ba
      go !n | n >= l    = v
            | otherwise = k (w n) (go (n + 1))
  in go 0

foldr' :: (Word8 -> a -> a) -> a -> ShortByteString -> a
foldr' k v = \sbs ->
  let l  = length sbs
      ba = asBA sbs
      w  = indexWord8Array ba
      go !n | n >= l    = v
            | otherwise = let !acc = go (n + 1)
                          in k (w n) acc
  in go 0
```

but doesn't that indicate that there's a problem with list fusion?