@hasufell 2022-07-04
Ok, so lazy `foldr` can also be sped up by marking both `unpack` and `unpackBytes` with `INLINE`. That suggests that `-fexpose-all-unfoldings` is not really enough and that unpack doesn't fuse at all (which seems pretty bad).

That only leaves strict `foldr'`. No combination of inlining had improved this, so I guess we can go with the manual version here.

That leaves that question whether we have other functions that don't fuse/inline properly as well.