@hasufell 2022-07-04
So, here are the benchmark results of the latest patchset, compared with the baseline: https://gist.github.com/hasufell/3710e86fcdfcd5f2c641e7d0a57e30ce

1. folds (strict and lazy), filter, unfoldrN are faster
2. raw unpack when just forcing the entire result is slower (kinda expected, since we have more thunks now?)
3. unpack fuses much better now though (e.g. `last . unpack` or `take 120 . unpack` are much faster)
4. added a lot of new `INLINE` pragmas (they had an effect)